==> notebooks/02_demographic.ipynb <==
# Config

# This script retrieves the list of organizations from the Open Data API of Barcelona.
# It uses the requests library to make API calls and pandas for data manipulation.
import requests
import os
import pandas as pd
from shapely import wkt

import numpy as np
import geopandas as gpd
from config import INTERIM_DATA_DIR, PROCESSED_DATA_DIR, CKAN_API_KEY, RAW_DATA_DIR
from io import StringIO

https://portaldades.ajuntament.barcelona.cat/ca/microdades/90dc3d6e-1c9a-4136-aed8-74bebb43e052

https://portaldades.ajuntament.barcelona.cat/ca/microdades/eea11ee9-cef5-41e8-841f-df2bbc809041

https://portaldades.ajuntament.barcelona.cat/ca/microdades/a45e1f19-5137-45bd-8979-645906fde55b

Revisar:
https://portaldades.ajuntament.barcelona.cat/ca/microdades/d8e40c96-9f1f-4fd3-86da-2baa1599616d

# Base API setPoblació de Barcelona agregada per edat en grups de cinc anys segons el registre del Padró Municipal d'Habitants a data 1 de gener de cada any
tings
API_URL = "https://portaldades.ajuntament.barcelona.cat/services/backend/rest/microdata/export?"
headers = {'X-IBM-Client-Id': CKAN_API_KEY}

# Define datasets (unique keys)
DATASETS = {
    'Vehicles per type': {
        'id': '90dc3d6e-1c9a-4136-aed8-74bebb43e052',
        'resource_id': '344e1547-8e13-4452-8fa8-f362601a9d52',
        'year': 2024
    },
    'Population per continent 2025': {
        'id': 'eea11ee9-cef5-41e8-841f-df2bbc809041',
        'resource_id': 'ab539ff9-7133-47ba-9114-51a6aa855d7d',
        'year': 2025
    },
    'Population per age group 2025': {
        'id': 'a45e1f19-5137-45bd-8979-645906fde55b',
        'resource_id': 'c5af1fec-95bc-4f25-8adc-c0313cfe0144',
        'year': 2025
    },
    'Rent per household': {
        'id': 'd8e40c96-9f1f-4fd3-86da-2baa1599616d',
        'resource_id': 'fd12fd1f-5fe6-4642-9ace-34503c2a9dd5',
        'year': 2023
    },
}

# List of JSON URLs
json_urls = [
    "https://opendata-ajuntament.barcelona.cat/resources/bcn/EstadisticaPadro/pad/2025/2025_pad_mdbas_nacionalitat-continent_sexe.json",
    "https://opendata-ajuntament.barcelona.cat/resources/bcn/EstadisticaPadro/pad/2025/2025_pad_mdbas_edat-q.json",
    "https://opendata-ajuntament.barcelona.cat/resources/bcn/EstadisticaUnitatsAdministratives/BarcelonaCiutat_SeccionsCensals.json"
]

# Load data

## Read CSV using the API


dfs = {}

for name, info in DATASETS.items():
    print(f"Fetching {name}...")
    params = {
        "id": info["id"],
        "idResource": info["resource_id"],
        "language": "ca",
        "fileformat": "CSV"
    }

    r = requests.get(API_URL, headers=headers, params=params)
    if r.status_code == 200:
        text = r.content.decode('utf-8')
        df = pd.read_csv(StringIO(text))
        dfs[name] = df
        print(f"  → Retrieved {len(df)} rows")
        df.to_csv(name.lower().replace(' ', '_') + ".csv", index=False)
    else:
        print(f"  ⚠️ Error {r.status_code}:")

print("\nExample preview:")
for k, v in dfs.items():
    print(f"\n{k}\n", v.head())

## Read JSON

def load_json_dataset(url):
    response = requests.get(url)
    if response.status_code != 200:
        raise Exception(f"Failed to fetch {url}: {response.status_code}")
    
    data = response.json()  # list of dictionaries
    df = pd.DataFrame(data)
    
    # Convert numeric columns if possible
    for col in ["Valor", "SEXE", "NACIONALITAT_CONTINENT", "EDAT_Q"]:
        if col in df.columns:
            df[col] = pd.to_numeric(df[col], errors="coerce")
    
    return df

# Load all JSONs
for url in json_urls:
    name = url.split("/")[-1].replace(".json", "")
    dfs[name] = load_json_dataset(url)
    print(f"{name} loaded with {len(dfs[name])} rows")

# Example preview
print(dfs["2025_pad_mdbas_nacionalitat-continent_sexe"].head())
print(dfs["2025_pad_mdbas_edat-q"].head())

## Read manually

# Rent
dfs["renda_bruta_llar"] = pd.read_csv(RAW_DATA_DIR/"2023_atles_renda_bruta_llar.csv")

# Data management (Code)

## Format census Tracks

#Get data
census_tracks = dfs["BarcelonaCiutat_SeccionsCensals"].copy()

#Clean geometries
census_tracks["geometria_etrs89"] = census_tracks["geometria_etrs89"].apply(wkt.loads)
census_tracks = gpd.GeoDataFrame(census_tracks, geometry="geometria_etrs89", crs="EPSG:25831")

#Correct codi
census_tracks["Seccio_Censal"] = census_tracks["codi_districte"].astype(str).str.zfill(2) + census_tracks["codi_seccio_censal"].astype(str).str.zfill(3)
census_tracks["Seccio_Censal"]

#Keep relevant columns
census_tracks = census_tracks[['Seccio_Censal', 'geometria_etrs89']]
census_tracks.rename(columns={"geometria_etrs89": "geometry"}, inplace=True)
census_tracks.set_geometry("geometry", inplace=True)
census_tracks.columns



census_tracks.head()

## Format data

# Vehicles per type

#Get data
vehicles_type = dfs["Vehicles per type"].copy()

#Correct codi
vehicles_type["Seccio_Censal"] = vehicles_type["Codi_Districte"].astype(str).str.zfill(2) + vehicles_type["Seccio_censal"].astype(str).str.zfill(3)

#Keep relevant columns
vehicles_type = vehicles_type[['Seccio_Censal', 'Tipologia_parc', 'Nombre_vehicles']]

# Make wide format

vehicles_type = vehicles_type.pivot_table(index=['Seccio_Censal'], columns='Tipologia_parc', values='Nombre_vehicles', fill_value=0, ).reset_index()
vehicles_type.columns.name = None  # remove the pivoted column name
vehicles_type.head(2)

# Vehicles per type

#Get data
rent = dfs["renda_bruta_llar"].copy()

#Correct codi
rent["Seccio_Censal"] = rent["Codi_Districte"].astype(str).str.zfill(2) + rent["Seccio_Censal"].astype(str).str.zfill(3)

# #Keep relevant columns
rent = rent[['Seccio_Censal', 'Import_Renda_Bruta_€']]

rent.head(10)

# Population by continent

#Get data
population = dfs["2025_pad_mdbas_nacionalitat-continent_sexe"].copy()

#Correct codi
population["Seccio_Censal"] = population["Seccio_Censal"].astype(str).str.zfill(5)

# #Keep relevant columns
population = population[['Seccio_Censal','Valor',	'NACIONALITAT_CONTINENT',	'SEXE']]

# # Make wide format
population_continent = population.pivot_table(index=['Seccio_Censal'], columns='NACIONALITAT_CONTINENT', values='Valor', fill_value=0, ).reset_index()
population_continent.columns.name = None  # remove the pivoted column name
population_continent.rename(columns={1:'Africa', 2:'America', 3:'Asia', 4:'Europe', 5:'Oceania'}, inplace=True)

population_gender = population.pivot_table(index=['Seccio_Censal'], columns='SEXE', values='Valor', fill_value=0, ).reset_index()
population_gender.columns.name = None  # remove the pivoted column name
population_gender.rename(columns={1:'Female', 2:'Male'}, inplace=True)

# population_continent.head(2)
population_gender.head()

#Population

# Population by continent

#Get data
age = dfs["2025_pad_mdbas_edat-q"].copy()

#Correct codi
age["Seccio_Censal"] = age["Seccio_Censal"].astype(str).str.zfill(5)

# #Keep relevant columns
age = age[['Seccio_Censal','Valor','EDAT_Q']]

# # Make wide format
age = age.pivot_table(index=['Seccio_Censal'], columns='EDAT_Q', values='Valor', fill_value=0, ).reset_index()
age.columns.name = None  # remove the pivoted column name
age.rename(columns={
    0: '<5 years',
    1: '5-9 years',
    2: '10-14 years',
    3: '15-19 years',
    4: '20-24 years',
    5: '25-29 years',
    6: '30-34 years',
    7: '35-39 years',
    8: '40-44 years',
    9: '45-49 years',
    10: '50-54 years',
    11: '55-59 years',
    12: '60-64 years',
    13: '65-69 years',
    14: '70-74 years',
    15: '75-79 years',
    16: '80-84 years',
    17: '85-89 years',
    18: '90-94 years',
    19: '95-99 years',
    20: '100+'
}, inplace=True)
age.head()

## Merge all data together

datasets = [census_tracks, vehicles_type, population_continent, population_gender, age, rent]
results = datasets[0]
for dataset in datasets[1:]:
    results = results.merge(dataset, on='Seccio_Censal', how='left')

# Plots

# Plots and visualizations

# Save results

#Save results and figures
results.to_parquet(INTERIM_DATA_DIR/"opendata.parquet")
