==> notebooks/05_cluster categories.ipynb <==
# Config

# Libraries
import planetary_computer
import pystac_client
import stackstac
import rasterio.plot
import matplotlib.pyplot as plt
import xarray as xr
import pandas as pd
import geopandas as gpd
from pathlib import Path 
from config import INTERIM_DATA_DIR, PROCESSED_DATA_DIR

from sentence_transformers import SentenceTransformer
from sklearn.cluster import KMeans
import numpy as np



# Params
CITY = "barcelona"
RES = 10


# Load data

result_wide = gpd.read_parquet(PROCESSED_DATA_DIR/f"{CITY}_h3_res{RES}_aggregated_wide.parquet")
print(result_wide.shape)
print(result_wide.columns)
result_wide.head()

# Data management (Code)

poi_types = result_points.columns.to_list()
poi_types.remove("h3_id")
poi_types.remove("geometry")


print(poi_types)

#Main body of code
model = SentenceTransformer('all-MiniLM-L6-v2')
embeddings = model.encode(list)

kmeans = KMeans(n_clusters=15, random_state=0)
labels = kmeans.fit_predict(embeddings)

clusters = {poi: label for poi, label in zip(poi_types, labels)}


cluster_names = {}
for cluster_id, pois in cluster_groups.items():
    indices = [poi_types.index(poi) for poi in pois]
    centroid = np.mean(embeddings[indices], axis=0)
    
    # Find closest POI to centroid
    distances = np.linalg.norm(embeddings[indices] - centroid, axis=1)
    name = pois[np.argmin(distances)]
    cluster_names[cluster_id] = name

cluster_names


# Plots

# Plots and visualizations

# Save results

#Save results and figures
